Refactor: service layer -> business layer 변경 및 ClubService의 getAllClub() 반환 값을 Club Domain으로 변경
